Refactor all Double parameters to the BigDecimal
- Refactor all classes that contains BigDecimal and/or operations with them

+++ src/main/kotlin/com/conrradocamacho/alugames/model/Recommendable.kt
@@ -1,7 +1,9 @@
 package com.conrradocamacho.alugames.model
 
+import java.math.BigDecimal
+
 interface Recommendable {
-    val average: Double
+    val average: BigDecimal
 
     fun recommend(grade: Int)
 }
